assignment_3/requirements.txt
MySQL-python==1.2.5
numpy==1.22.4
pandas==1.4.3
Pillow==9.2.0
SQLAlchemy==1.4.39
streamlit==1.12.0
